# docs/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "GrownupsVet API",
    "description" : "Contrato generado desde código para acceso, perfil, sesiones, mascotas, recuperación, personal con invitaciones y disponibilidad veterinaria. El correo requiere configuración SMTP. La URL de activación del portal contiene un placeholder configurable. Reservas, consulta automática de COMVEZCOL y plantillas Thymeleaf quedan fuera de este incremento.",
    "version" : "0.6.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/veterinarians/{veterinarianId}/qualifications/{qualificationId}" : {
      "get" : {
        "tags" : [ "Títulos y diplomas veterinarios" ],
        "summary" : "Consultar un título académico",
        "operationId" : "getVeterinarianQualification",
        "parameters" : [ {
          "name" : "veterinarianId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "qualificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Petición o parámetros inválidos",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Token ausente o inválido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Rol sin permiso para esta operación",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Recurso inexistente o no accesible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "Correo, matrícula o estado en conflicto",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "406" : {
            "description" : "Formato de respuesta no disponible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "415" : {
            "description" : "Tipo de contenido no admitido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Fallo interno sin información sensible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "200" : {
            "description" : "Título académico",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VeterinarianQualificationResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Títulos y diplomas veterinarios" ],
        "summary" : "Reemplazar datos del título",
        "description" : "Cambiar datos académicos retira la publicación del diploma para permitir una nueva revisión; conserva el archivo privado. El título base mantiene el tipo UNDERGRADUATE.",
        "operationId" : "updateVeterinarianQualification",
        "parameters" : [ {
          "name" : "veterinarianId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "qualificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VeterinarianQualificationRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Petición o parámetros inválidos",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Token ausente o inválido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Rol sin permiso para esta operación",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Recurso inexistente o no accesible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "Correo, matrícula o estado en conflicto",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "406" : {
            "description" : "Formato de respuesta no disponible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "415" : {
            "description" : "Tipo de contenido no admitido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Fallo interno sin información sensible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "200" : {
            "description" : "Título actualizado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VeterinarianQualificationResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Títulos y diplomas veterinarios" ],
        "summary" : "Eliminar título adicional y su diploma",
        "description" : "El título base es obligatorio y no se puede eliminar.",
        "operationId" : "deleteVeterinarianQualification",
        "parameters" : [ {
          "name" : "veterinarianId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "qualificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Petición o parámetros inválidos",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Token ausente o inválido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Rol sin permiso para esta operación",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Recurso inexistente o no accesible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "Correo, matrícula o estado en conflicto",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "406" : {
            "description" : "Formato de respuesta no disponible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "415" : {
            "description" : "Tipo de contenido no admitido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Fallo interno sin información sensible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "204" : {
            "description" : "Título eliminado"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/veterinarians/{id}" : {
      "get" : {
        "tags" : [ "Administración de veterinarios" ],
        "summary" : "Consultar veterinario",
        "operationId" : "getVeterinarian",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Petición o parámetros inválidos",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Token ausente o inválido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Rol sin permiso para esta operación",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Recurso inexistente o no accesible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "Correo, matrícula o estado en conflicto",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "406" : {
            "description" : "Formato de respuesta no disponible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "415" : {
            "description" : "Tipo de contenido no admitido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Fallo interno sin información sensible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "200" : {
            "description" : "Perfil administrativo",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VeterinarianResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Administración de veterinarios" ],
        "summary" : "Reemplazar datos editables del perfil",
        "description" : "Reemplazo completo de los datos editables. Conserva correo, rol y estado. Modificar datos académicos retira la publicación de su diploma hasta una nueva revisión.",
        "operationId" : "updateVeterinarian",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateVeterinarianRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Petición o parámetros inválidos",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Token ausente o inválido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Rol sin permiso para esta operación",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Recurso inexistente o no accesible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "Correo, matrícula o estado en conflicto",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "406" : {
            "description" : "Formato de respuesta no disponible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "415" : {
            "description" : "Tipo de contenido no admitido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Fallo interno sin información sensible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "200" : {
            "description" : "Perfil actualizado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VeterinarianResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "VeterinarianQualificationRequestDTO" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "UNDERGRADUATE", "SPECIALIZATION", "MASTERS", "DOCTORATE" ]
          },
          "title" : {
            "type" : "string",
            "example" : "Especialización en Medicina Interna",
            "maxLength" : 200,
            "minLength" : 0
          },
          "institution" : {
            "type" : "string",
            "example" : "Universidad del Quindío",
            "maxLength" : 200,
            "minLength" : 0
          },
          "graduationYear" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "Opcional. No puede ser posterior al año actual en America/Bogota.",
            "maximum" : 9999,
            "minimum" : 1900
          }
        },
        "required" : [ "institution", "title", "type" ]
      },
      "ApiErrorResponseDTO" : {
        "type" : "object",
        "description" : "Respuesta de error RFC 9457, con código funcional y errores por campo.",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "Referencia URI al tipo de problema.",
            "example" : "about:blank"
          },
          "title" : {
            "type" : "string",
            "description" : "Resumen del problema.",
            "example" : "Solicitud incorrecta"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Estado HTTP de la respuesta.",
            "example" : 400,
            "maximum" : 599,
            "minimum" : 400
          },
          "detail" : {
            "type" : "string",
            "description" : "Explicación de lo ocurrido en esta solicitud.",
            "example" : "Revisa los campos indicados."
          },
          "instance" : {
            "type" : "string",
            "format" : "uri-reference",
            "description" : "Referencia URI opcional que identifica esta ocurrencia del problema.",
            "example" : "urn:uuid:7d667530-867b-4c78-a458-0904fb82d574"
          },
          "errorCode" : {
            "type" : "string",
            "description" : "Código funcional estable; el frontend no debe comparar los mensajes.",
            "example" : "INVALID_REQUEST"
          },
          "fieldErrors" : {
            "type" : "array",
            "description" : "Errores por campo. Es una lista vacía cuando el error no corresponde a campos específicos.",
            "items" : {
              "$ref" : "#/components/schemas/FieldValidationErrorResponseDTO"
            }
          }
        },
        "required" : [ "detail", "errorCode", "fieldErrors", "status", "title", "type" ]
      },
...
      "VeterinarianQualificationResponseDTO" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "UNDERGRADUATE", "SPECIALIZATION", "MASTERS", "DOCTORATE" ]
          },
          "title" : {
            "type" : "string",
            "maxLength" : 200
          },
          "institution" : {
            "type" : "string",
            "maxLength" : 200
          },
          "graduationYear" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "maximum" : 9999,
            "minimum" : 1900
          },
          "isBaseDegree" : {
            "type" : "boolean"
          },
          "diplomaAvailable" : {
            "type" : "boolean"
          },
          "diplomaPublished" : {
            "type" : "boolean"
          }
        },
        "required" : [ "diplomaAvailable", "diplomaPublished", "graduationYear", "id", "institution", "isBaseDegree", "title", "type" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "JWT RS256 supplied in the Authorization header. Tokens expire 24 hours after login.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}